requirements.txt
transformers
requests
Pillow
torch
torchvision
flask
huggingface_hub[cli]
